+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Document derivation of offsets
 - GH release publish to pypi
 - zlim handling and consistent depth
+- Check AxesAnaglyph.imshow_stereo() non-redcyan coloring
 
 ----